+++ tests/unmanaged.rs
@@ -1,11 +1,15 @@
 #[cfg(feature = "unmanaged")]
 mod tests {
 
+    use std::time::Duration;
+
+    use tokio::time::{timeout, interval};
+
     use deadpool::unmanaged::Pool;
 
     #[tokio::test]
     async fn test_unmanaged_basic() {
-        let pool = Pool::new(vec![(), (), ()]);
+        let pool = Pool::from(vec![(), (), ()]);
 
         let status = pool.status();
         assert_eq!(status.size, 3);
@@ -32,7 +36,7 @@ mod tests {
 
     #[tokio::test(threaded_scheduler)]
     async fn test_unmanaged_concurrent() {
-        let pool = Pool::new(vec![0usize, 0, 0]);
+        let pool = Pool::from(vec![0usize, 0, 0]);
 
         // Spawn tasks
         let futures = (0..100)
@@ -57,5 +61,57 @@ mod tests {
         let values = [pool.get().await, pool.get().await, pool.get().await];
 
         assert_eq!(values.iter().map(|obj| **obj).sum::<usize>(), 100);
+    }
+
+    #[tokio::test(threaded_scheduler)]
+    async fn test_unmanaged_add_remove() {
+        let pool = Pool::new(2);
+        pool.add(1).await;
+        assert_eq!(pool.status().size, 1);
+        pool.add(2).await;
+        assert_eq!(pool.status().size, 2);
+        assert!(timeout(Duration::from_millis(10), pool.add(3)).await.is_err(), "adding a third item should timeout");
+        pool.remove().await;
+        assert_eq!(pool.status().size, 1);
+        assert!(timeout(Duration::from_millis(10), pool.add(3)).await.is_ok(), "adding a third item should not timeout");
+        pool.remove().await;
+        assert_eq!(pool.status().size, 1);
+        pool.remove().await;
+        assert_eq!(pool.status().size, 0);
+    }
+
+    #[tokio::test(threaded_scheduler)]
+    async fn test_unmanaged_try_add_try_remove() {
+        let pool = Pool::new(2);
+        pool.try_add(1).unwrap();
+        assert_eq!(pool.status().size, 1);
+        pool.try_add(2).unwrap();
+        assert_eq!(pool.status().size, 2);
+        assert!(pool.try_add(3).is_err());
+        pool.try_remove().unwrap();
+        assert_eq!(pool.status().size, 1);
+        assert!(pool.try_add(3).is_ok());
+        pool.try_remove().unwrap();
+        assert_eq!(pool.status().size, 1);
+        pool.try_remove().unwrap();
+        assert_eq!(pool.status().size, 0);
+    }
+
+    #[tokio::test(threaded_scheduler)]
+    async fn test_unmanaged_add_timeout() {
+        let pool = Pool::from(vec![1]);
+        let add = {
+            let pool = pool.clone();
+            tokio::spawn(async move {
+                pool.add(2).await;
+            })
+        };
+        let mut iv = interval(Duration::from_millis(10));
+        iv.tick().await;
+        iv.tick().await;
+        pool.try_remove().unwrap();
+        assert!(timeout(Duration::from_millis(10), add).await.is_ok(), "add should not timeout");
+        assert_eq!(pool.status().size, 1);
+        assert_eq!(pool.try_remove().unwrap(), 2);
     }
 }